requirements.txt
streamlit~=1.35.0
requests~=2.32.2
matplotlib~=3.9.0
plotly~=5.22.0
groq~=0.8.0
pandas~=2.2.2
python-dotenv~=1.0.1